Create ingredient spoons

+++ Source/Model/NewAddIngredientAmount/NewAddIngredientAmountType.swift
@@ -8,6 +8,6 @@ struct NewAddIngredientAmountType
         static let defaultQuantity:Float = 1
         static let defaultGrams:Int32 = 1
         static let defaultMililitres:Int32 = 1
-        static let defaultSpoons:Float = 1
+        static let defaultSpoons:Int16 = 1
     }
 }
